requirements.txt
fastapi
uvicorn[standard]
httpx
playwright
pandas
pdfplumber
python-multipart
pydantic
matplotlib
openpyxl
python-dotenv
requests
